Linked program to WebGLRenderingContext # Please enter the commit message for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit. # # On branch master # Changes to be committed: #	modified:   app.js # # Untracked files: #	shaders.js # I'm not entirely sure if the title of this commit is right. From what I understood, linking a program is the final step in the process of preparing information that the GPU can understand. `linkProgram()` is a WebGLRenderingContext function. It takes in a WebGLProgram, which we have already created and attached shaders to.
I think the important take-away here is, linking the program, means that we're done with a level of abstraction. From here on out, we do not need to manage both the vertex shader and the fragment shader as separate things. This is not to say that vertex shaders and fragment shaders won't be important as concepts later on, they still exist but we can access them through the newly created program.

I claim this as the end of the era of shaders.

Not that we ever necessarily entered the era of shaders. I guess that's the thing with living through an era. You don't know what era you're currently in until someone in the future labels it. I don't think people in the dark ages *knew* they were in the dark ages.